test: schedule test logins on correct day
As the current time is moving on throught the test, the calculation of "n days
ago" ends up picking a time which is slightly more than "n days ago", so reduce
n by one. This error was mostly neutralised by the rotation threshold throwing
away microseconds, but with that fixed, the tests always fail unless this is
also fixed.