import: match login by tty if non-zero pid does not match
The existing code matches a utmp logout entry to a login entry by the preferred
pid method if possible and then by tty if the pid is zero. This does not
correspond to some servers that record the logout with a different value in the
pid field, resulting in the logout times being recorded as 'crash'.