Use empty memory table instead of failing to read empty file
Previously, attempting to read an empty file gave an error
message and non-zero return code. Instead, when asked to open
an empty file as a read-only database, open a memory database
and populate it with an empty table. This avoids needing any
special case handling in calling code and matches the behaviour
of classic 'last' on an empty file.