last -x: show shutdown entries before reboot ones
Move the calculation and printing of the shutdown entry before the printing of
the reboot entry. This requires using different storage elements for this tuple
of formatted times; use a struct for this to avoid expressing the relevant
dimensions in more places in the code, inventing new variable names or
triggering -Wshadow.